Only perform ivy configuration selection on ivy components
When an ivy component is selecting from another component without variants, it tries to perform
selection as if the target component was an ivy component. However, this can attempt to select
against a local componenta as an ivy component if that local component does not declare any
consumable configurations with attributes.

In order to simplify component state/metadata models to not need to advertise configurations,
we restrict ivy legacy selection to only take place against ivy components. This way, we will
eventually be able to remove the getConfiguration method from all non-ivy components.