Clarify that lazy state should live in a State type
Much of the locking, caching, etc is done twice. The intention is for the metadata to be immutable and should not run user code.
The State type should be responsible for calculating and calculating values for use in graph analysis.
Right now, the Metadata types do a lot of this work. They should only contain static data.