Avoid parent component reference when calculating artifacts
We previously relied on the parent component when calculating configuration artifacts in order
to get the artifacts of other configurations in the component. We should deprecate this behavior
of inheriting artifacts, but in the meantime, we avoid the parent reference by re-calculating
the artifact instances when we do inherit artifacts.

This is a bit inefficient, but should not affect performance. As a result, we get to avoid the ugly
circular parent reference but also get to avoid the getConfiguration call on the component. We want
to remove this method, as we want to get rid of the configuration concept from the resolution engine.