Do not allow requests for arbitrary local configuration metadata/state
We enhance getVariantsForGraphTraversal to include all consumable variants, even those without attributes,
allowing us to replace the getConfiguration method with a getRootVariant method, which is only called
when performing graph resolution.

This allows us to avoid calling getConfiguration when building variants for attribute matching, simplifying the
process. This way, when we add proper locking to these types we do not need to lock over multiple data structures.

As a side-effect of this, we will create duplicate metadatas/states if a configuration is both resolved and consumed in different
graph resolutions. This is minor as this situation is quite rare, and already leads to deprecation warnings.

Furthermore, this change also makes it more difficult for the metadata engine to know if a configuration does not exist or is it
simply non consumable when a user requests a configuration by name and none is found. This is not a big deal, since we enhance the
variant-missing message to mention that you may have requested a non-consumable configuration.