Decouple version conflict error handling
This commit decouples the resolution path from needing to know about a Configuration's name and path
when handling version conflict failures. This data will not always be available for all resolutions,
since in future changes we will add the ability to resolve things that are not configurations.

The primary change here adds a FailureContext to a ResovleContext, which can provide context-specific
resolutions for failure scenarios. Specifically, we add a method to provide failure resolutions that
instruct the user to call the dependencyInsight task on the configuration being resolved.

Additionally, we perform some cleanup in the exception mapping logic. Specifically, we pull out
DefaultLenientConfiguration.ArtifactResolveException into a concrete TypedResolveException
type and update the code throughout to use this subtype. Further, we deprecate public
constructors on ResolveException, as users should not be instantiating that type.