Allow detached configurations to resolve current project
Previously, the root component of a detached configuration would claim to have the same gav as the project
that created the detached configuration. This caused the resolution engine to treat the root component of
the detached configuration as an instance of the project module.

However, the detached configuration's root component only has the detached configuration and none of the
variants of the project. So, to resolve this we attach a suffix to the module ID of the root component
of the detached configuration to ensure it does not conflict with the project.

This can be flimsy in certain situations. If there was another project with the same name as the constructed
module ID (with the suffix) then the detached configuration's root component will conflict with that component.
However, this is likely rare, as the constucted module ID would look something like org:module-detachedConfiguration1:version
or org:module-runtimeClasspathCopy:version.

Future work could aim to improve this by adding a ComponentIdentifier-like counterpart to ModuleIdentifier. Then,
we could allow the root component here to have some opaque module idientifer that does not collide with any named
module identifier